Fix communication error handling

+++ src/ypspur_ros.cpp
@@ -742,7 +742,8 @@ public:
   }
   ~YpspurRosNode()
   {
-    if (pid_ > 0)
+    // Kill ypspur-coordinator if the communication is still active.
+    if (pid_ > 0 && YP::YP_get_error_state() == 0)
     {
       ROS_INFO("killing ypspur-coordinator (%d)", (int)pid_);
       kill(pid_, SIGINT);
@@ -750,7 +751,6 @@ public:
       waitpid(pid_, &status, 0);
       ROS_INFO("ypspur-coordinator is killed (status: %d)", status);
     }
-    ros::shutdown();
   }
   void spin()
   {
@@ -819,8 +819,8 @@ public:
         if (!simulate_control_)
         {
           t = YP::YPSpur_get_pos(YP::CS_BS, &x, &y, &yaw);
-          if (t == 0.0)
-            t = now.toSec();
+          if (t <= 0.0)
+            break;
           YP::YPSpur_get_vel(&v, &w);
         }
         else
@@ -854,8 +854,8 @@ public:
         tf_broadcaster_.sendTransform(odom_trans);
 
         t = YP::YPSpur_get_force(&wrench.wrench.force.x, &wrench.wrench.torque.z);
-        if (t == 0.0)
-          t = now.toSec();
+        if (t <= 0.0)
+          break;
         wrench.header.stamp = ros::Time(t);
         wrench.wrench.force.y = 0;
         wrench.wrench.force.z = 0;
@@ -917,9 +917,9 @@ public:
               }
               i++;
             }
-            if (t == 0.0)
-              t = ros::Time::now().toSec();
           }
+          if (t <= 0.0)
+            break;
           joint.header.stamp = ros::Time(t);
         }
         else
@@ -1192,10 +1192,8 @@ public:
       updateDiagnostics(now);
 
       if (YP::YP_get_error_state())
-      {
-        ROS_ERROR("ypspur-coordinator is not active");
         break;
-      }
+
       ros::spinOnce();
       loop.sleep();
 
@@ -1222,6 +1220,9 @@ public:
         break;
       }
     }
+    if (YP::YP_get_error_state())
+      ROS_ERROR("ypspur-coordinator is not active");
+
     ROS_INFO("ypspur_ros main loop terminated");
   }
 };
